+++ frontend/src/components/icons/Icons.tsx
@@ -6,6 +6,7 @@ import { ReactComponent as LogoRound } from "../../assets/icons/logo_round.svg";
 import { ReactComponent as Star } from "../../assets/icons/star.svg";
 import { ReactComponent as Transaction } from "../../assets/icons/transaction.svg";
 import { ReactComponent as ArtistPalette } from "../../assets/icons/artist_palette.svg";
+import { ReactComponent as Settings } from "../../assets/icons/settings.svg";
 import { ReactComponent as CircleArrowLeft } from "../../assets/icons/circle_arrow_left.svg";
 import { ReactComponent as Plus } from "../../assets/icons/plus.svg";
 import { ReactComponent as Close } from "../../assets/icons/close.svg";
@@ -15,6 +16,8 @@ import { ReactComponent as Cancel } from "../../assets/icons/cancel.svg";
 import { ReactComponent as Link } from "../../assets/icons/link.svg";
 import { ReactComponent as Script } from "../../assets/icons/script.svg";
 import { ReactComponent as QuestionMark } from "../../assets/icons/question-mark.svg";
+import { ReactComponent as Switch } from "../../assets/icons/switch.svg";
+import { ReactComponent as Restart } from "../../assets/icons/restart.svg";
 
 export const FlowserIcon = {
   Account,
@@ -27,6 +30,9 @@ export const FlowserIcon = {
   Script,
   QuestionMark,
   ArtistPalette,
+  Settings,
+  Switch,
+  Restart,
   CircleArrowLeft,
   Plus,
   Close,
