+++ frontend/src/components/process/ManagedProcess.module.scss
@@ -1,0 +1,55 @@
+@import "styles/colors";
+@import "styles/spacings";
+
+.processItem {
+  justify-content: space-between;
+
+  .label {
+    font-size: $font-size-normal;
+  }
+
+  .processItemActions {
+    display: flex;
+    align-items: center;
+
+    .restartButton {
+      width: 15px;
+      height: 15px;
+      margin-right: $spacing-base;
+
+      &:hover {
+        svg {
+          transform: scale(1.1) rotate(40deg);
+        }
+      }
+
+      svg {
+        transition: 0.3s ease-in-out;
+        width: 100%;
+        height: 100%;
+
+        path {
+          fill: $grey;
+        }
+      }
+    }
+  }
+}
+
+.processStatus {
+  border-radius: 50%;
+  width: 12px;
+  height: 12px;
+}
+
+.processStatusError {
+  background: $color-red-light;
+}
+
+.processStatusRunning {
+  background: $strong-green;
+}
+
+.processStatusNotRunning {
+  background: $grey;
+}
