+++ frontend/src/components/side-navigation/SideNavigation.tsx
@@ -5,12 +5,18 @@ import { NavLink } from "react-router-dom";
 import { FlowserIcon } from "components/icons/Icons";
 import { SizedBox } from "../sized-box/SizedBox";
 import classNames from "classnames";
+import { useGetCurrentProject } from "../../hooks/use-api";
+import { useProjectActions } from "../../contexts/project.context";
 
 type SideNavigationProps = {
   className?: string;
 };
 
 export function SideNavigation(props: SideNavigationProps): ReactElement {
+  const { switchProject } = useProjectActions();
+  const { data: currentProjectData } = useGetCurrentProject();
+  const { project: currentProject } = currentProjectData ?? {};
+
   return (
     <div className={classNames(classes.root, props.className)}>
       <FlowserLogo />
@@ -22,6 +28,17 @@ export function SideNavigation(props: SideNavigationProps): ReactElement {
       <ProjectLink to={routes.events} icon={FlowserIcon.Star} />
       <ProjectLink to={routes.interactions} icon={FlowserIcon.CursorClick} />
       <ProjectLink to={routes.project} icon={FlowserIcon.ArtistPalette} />
+      {currentProject && (
+        <ProjectLink
+          to={`/start/configure/${currentProject.id}`}
+          icon={FlowserIcon.Settings}
+        />
+      )}
+      <ProjectLink
+        to={routes.start}
+        icon={FlowserIcon.Switch}
+        onClick={switchProject}
+      />
     </div>
   );
 }
@@ -34,6 +51,7 @@ function FlowserLogo() {
 function ProjectLink(props: {
   to: string;
   icon: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
+  onClick?: () => void;
 }) {
   const Icon = props.icon;
   const iconSize = 20;
@@ -42,6 +60,7 @@ function ProjectLink(props: {
       to={props.to}
       className={classes.inactiveLink}
       activeClassName={classes.activeLink}
+      onClick={props.onClick}
     >
       <Icon width={iconSize} height={iconSize} />
     </NavLink>
